requirements.txt
fastapi
uvicorn
jinja2
asyncpg
python-dotenv
plotly
numpy
pandas

